Convert some floating point math to better fixed point math
This math was leading to instability across seeds depending on the
particular way in which floating point math was done on a build and
system.

This uses a suggestion from |amethyst - since callers always use
`percent` values that are divisible by 10, can just do the math with a
single decimal point rather than two, which takes up a lot less space.
This makes the calculation more accurate than what it replaces (though
it still truncates, rather than rounds).  For that reason, I have put in
a temporary stopgap with the quirks of the previous implementation when
run using the SSE version of the old floating point calculation. I will
remove this at some point after 0.23.2 is tagged.  (But for now, this
keeps the gotm stable.)