Add a menu shortcut to select last unequipped item
In equipment-related menus and prompts, make ';' (semicolon) an alias
for the letter of the player's most recently unequipped item. The goal
of this change is to reduce the amount of time the player spends
scanning through menus to remember which item is where. Examples:

* Wear-IDing a good cloak ('f') and dropping your old one: Wfd;<Enter>
* Wear-IDing a bad cloak ('f') and re-equipping the old: WfW;d;<Enter>
* Swapping to a utility ring ('g') for a certain enemy, then going back
to your usual ring: Pg<P;<

In those examples, we saved a little mental effort by not having to keep
track of the location in inventory of the old cloak and ring.

Notes/possible issues:

The semicolon key is losing existing, though undocumented, behavior as
one of several ways send a page-up key (besides '<', PGUP itself, and
inconsistently '-').

The fact that this alias is usable on the (d)rop menu is documented both
in crawl_manual.txt and on the d_ help screen, but for i, w, W, and P
it's manual-only. (Those other help screens don't yet exist.)