Better newgame messaging when a game already exists
Before, if a game already existed under the name, any menu option that
would start a game with that name prefilled (from prefs) would just load
the existing game, regardless of game type. Before seeded games, this
only affected tutorial vs. regular games, but it's now even more
confusing with seeded games. These menus are still a bit of a mess, so
this could use some more work (it might be better to prompt for a
different name from the seed menu, or do the check before opening that
menu).