Don't place shafts in ebering_ghost_disaster_area (11926)
This commit prevents the vaults ebering_ghost_disaster_area and
ebering_vaults_ghost_disaster_area from placing shaft traps, instead
placing only water and lava tiles, implementing option 2 in 11926.
This change is to prevent ghosts in these vaults from being shafted
immediately when they enter line of sight of the player, before the
player can do anything to stop this.

These vaults were introduced by commits bb431e2 and a6a7277
respectively, and appear to have been introduced with the idea that
ghosts cannot be shafted. Ghosts can now be shafted, as of 8c2e593, and
this means that these vaults in particular could cause ghosts to be
shafted to lower floors, leaving behind an empty (or mostly empty) ghost
vault for the player to raid, and allowing the ghost to be encountered
outside of its vault on a lower floor, with the player then unable to
choose whether to fight the ghost.