Don't use a more() for LOS spells (kitchen_ace)
Currently LOS attack spells use a more() prompt every time they fire.
This may be so that the flash animation they use can actually have time
to display. The more() prompt is not removable by the player and is
especially annoying for frequent LOS spells under e.g. the Singing
Sword. This commit removes the more() and instead uses a delayed flash
animation of 300ms, like we do for many other LOS-based spells.