Always display the duration of the last action on the HUD (Yermak)
Because of how the score is calculated, it might be more useful to have
player turns instead of elapsed time on the HUD.
`show_game_time = false` replaces time with player turns, but also
hides the duration of the last action. The HUD is the only place that
shows this duration, so players who care about their score have to use
the default value for show_game_time and have to regularly check their
turns via the E or % commands.

After this commit, the HUD will display the duration regardless of the
show_game_time's value.