Better separate level generation from level loading/entry
These were previously completely intertwined. Before 0.23, level
generation was only ever done on an attempt to enter a new level. For
0.23, level generation was a special case load mode for the call to
load_level, with complicated shared setup code for the generation call.
This commit pulls the relevant setup and cleanup out of load_level so
that a level can be generated without needing to call load_level at all.
There is a small cost in code duplication, but the comprehension cost
for doing this together with level loading was much higher.