Properly delete visited marker when a level is deleted
and also clarify the semantics of level_visited. This fixes weird
behavior when revisiting pan levels (it will also affect zigs, and
revisiting portal vaults in wizmode) where state was not correctly
initialized when entering a newly generated level, because the player
had visited a level with that level_id before.

If there's any games still saved in pan with the broken state, after
upgrading to this they can fix with a save and reload.