Make StashTrack pay attention to traps
When updating stashes in LOS, take traps seriously enough as a feature
to add a stash for them.

Something like this may have been already intended, since stashes have a
trap field that is supposed to be set to a non-trivial value in
Stash::update. This part of the code wasn't reached since
is_boring_feature(feat) is false when feat is a trap. (See also commits
23c5277317de657f8376d1b5f33144948a811a42 and
cbc0dd1c0e73770878e9a6df1b455e57995ef921.)

This change has traps added to the stash tracker as the player comes
upon them. Unfortunately players of ongoing games will not be able to
search for traps they have encountered that are no longer in LOS.