Make stash searching for a place abbreviation pull up features
Prior to this change, searching stashes with the search term '.' while
on D:2 and searching for 'D:2' while on D:2 had different results. This
was because the search term was matched against a location prefix for
items but not for features.

According to the search help screen the search term '.' should list
everything on the current level, so I conclude the difference here is
unintentional.