Don't place traps under vault placed items and monsters (11983)
Randomly placed monsters and items do not place on traps, however the
dungeon builder was placing random traps under vault placed items and
monsters occasionally. This led to runes on dispersal traps as the worst
outcome, but the asymmetry was occasionally noticed for other trap types
too. Vaults that wish to have monsters or items start on a trapped grid
will need to request this explicitly.