Clean up and improve gozag gold detection
This consolidates a bunch of gozag code and removes an ASSERT that I
think was over-eager, firing when update_item_at doesn't lead to a gold
pile in the stash tracker. There had previously been two versions of
this code, and only one had the check, but it doesn't seem so obvious to
me that update_item_at should really be expected to make any arbitrary
gold pile visible in the stash.

This commit also makes gold detection work in the abyss beyond the
initial levelgen.