Test a theory about hep abyss crashes
This is a conceivable way for hep ancestor items to get copied to the
grid, if somehow the ancestor becomes invalid before this call but
cleanup of the ancestor hasn't happened yet (since held_by_monster will
return false if the monster that was holding the item isn't valid). If
it trips, this ASSERT will give a more direct line on what the setup
conditions for this situation might be, which I still don't understand.