Mention draining gaze's healing effect
Draining gaze heals ghost moths and eyes of draining for the amount of
MP drained by the gaze. This commit adds a note to this effect to the
draining gaze spell description.

Note: orange crystal statues also cast this spell, but they don't get
healed from it because they cannot regenerate HP.