Don't target chain spell arcs through actors (12003)
In the target selection loop in cast_chain_spell, no check was made to ensure
that the arc from source to target was not obstructed by an actor. As a result,
the special-cased treatment of the caster was ignored occasionally. By tracing
the ray for each potential target and checking for obstructions this problem is
avoided.