Don't let monster digging allow out-of-los casting (11945, 11969)
Since 5b187de83c, monsters that cast dig do so on a movement action (the
burrowing mechanism), not usually as a regular spell. In fact, I
couldn't get a monster to ever cast dig normally. Therefore, no monsters
ever need to cast dig out-of-los via the normal spell code, and they
definitely don't need to be able to cast arbitrary spells out-of-los if
they have dig.

I *think* that this is a pretty old bug, and only used to trigger under
fairly rare conditions; a cacodemon still couldn't cast energy beam
through a wall (though it would try). However, it would work through fog
as reported in 11945. What changed recently is that LRD no longer
destroys walls but is still area targeted, so that in some
configurations a caster could attack out of los:

>   #
>  @#2
> #

A LRD caster (e.g. a cacodemon) in this position could succesfully cast
LRD on the wall in the lower left, damaging the player, instead of
moving. In my testing a direct LRD cast, e.g. at the wall between the
player, would still cast but have no effect or messaging, so even this
case would still involve some monster wasted turns. Before the LRD
change I think the caster could cast LRD succesfully on this square,
destroying it in the process, so the setup conditions for the bug were
much less common.  Hence the recent reports of this bug.