Optimize find_maps_for_tag somewhat
This gets about 10x speedup on overflow temple selection (on a
non-optimized build). It's still not great (since the overflow selection
code is pretty insane, could be as many as 300 calls to
find_maps_for_tag in my testing), but now according to my profiling the
bottleneck isn't the use of set<vector>, but rather the construction of
an iterator for each has_tag.

It's worth noting that 99% of the calls to map_def::has_tag in the
codebase have a single tag, and don't need this complexity. However, for
typical uses I didn't see any obvious gain from optimizing them.