Rewrite overflow vault selection code
The old version brute-force searched through every 2-3 god combination
among the gods selected for overflow vaults, with a somewhat indirect
and complicated stopping criteria (with a goto, even) and a call to
`find_maps_for_tag` for every combination it reached before the stopping
condition. This is a huge search space and even with the stopping
criteria it would still consider hundreds to thousands of combinations.
On the other hand there are less than 20 vaults *total* that this code
could even find.

This rewrite flips things, and searches through the temple_overflow_n
vaults themselves for compatible vaults rather than the space of god
combinations. It's *much* faster.

I expect this will change some seeds, since the order of rolls before
was determined by bitvector order, and this will be determined by vault
list order; also, where there are multiple vaults with the same god tags
the weight calculation is different in this implementation. However, in
my testing the results have been basically the same for most seeds. This
code probably is biased towards 2-altar vaults over 3-altar vaults as
is.