Cache the top three vault tags as bools
These tags account for >50% of the `has_tag` calls, which over the
course of a full pregeneration run really adds up -- this is nearly
5000000 calls for `has_tag`. This change shaves about 0.5s off a full
(non-optimized) pregen run.

This commit also adds some (disabled) profiling stuff for tags. For
reference here's the call count for the top 7 tags, everything lower
than this is <1%:

  208774 ( 2%): vaults_empty
  208774 ( 2%): vaults_room
  439758 ( 5%): layout
  550983 ( 6%): place_unique
 1150916 (14%): overwritable
 1243580 (15%): extra
 2554449 (31%): minivault
Total: 8144269