Don't let locale-specific casing affect ascii characters
If a locale changes the upper/lower case correspondents of an ascii
character, this would break many things. For example, as in #846, in
Turkish (tr_TR.utf8), the lowercase version of I is a dotless i (ı).
Since these case functions are used for all sorts of internal purposes
(e.g. producing filenames from capitalized tile names), crawl would
fail to even start. Though this issue only illustrated the first error,
this is the tip of the iceberg and nearly every use of lowercasing was
broken in this locale.

I think, ideally, we'd be using two different case swapping functions:
one for internal data (which is always ASCII) and one for UI-facing
strings (which may be in arbitrary UTF-8 encodings). However, when I
tried to adjust the API along these lines to fix Turkish, it turned out
that there is not a clean distinction in crawlcode between these two
uses and even stuff like the name functions which might be expected to
be UI-facing are used for internal purposes as well. Based on my
attempts, if this does come up for what little localization we have, it
would be better/safer to let the ASCII behavior be the default for ascii
characters and add in localized casing piecemeal.

This commit also adjusts various hotkeys, and I'm not sure what impact
this will have (it might be going too far); it could cause problems to
e.g. users on a Turkish keyboard.