Let X be zoomed differently from main map, default to 60%
In console, the `X` view is really useful in terms of getting a quick
overview of the level map. However, it's always struck me as a bit
useless / vestigial on tiles, because the amount of map displayed wasn't
much different from what `x` displays, just with a different set of keys
and behavior. The situation is slightly better on webtiles, where at
least the message pane is hidden, but on local tiles it literally was
not different.

This commit allows the `X` view to be zoomed differently from the main
map, and defaults it to 60% zoom (as opposed to default 100% zoom). You
can also now change the zoom in game using {,}. On local tiles, ctrl-'-'
and ctrl-'=' also work (as with the main view). To return this behavior
to its previous state, set the (new) rc option `tile_map_scale` to 1.0.

There are a few awkward issues remaining. In webtiles, some status
messages printed using `mpr` are no longer visible in `X` mode, though
prompts will work (see 65f37df). I'm not yet sure what the best solution
for these is. Also, unlike local tiles, webtiles lets you zoom beyond
what will actually display, because I couldn't figure out a simple way
of getting this information back from the client (though there are
certainly complex ways of doing this).